(WIP) iterative search implemented (not debugged)

+++ include/candidate.hpp
@@ -23,7 +23,7 @@ void bestScoreKernel(const uchar* F, double* S, int* D,
                      int width, int height);
 
 __global__
-void candidateThresholdKernel(const double *S, const uchar *D, uchar *C,
+void candidateThresholdKernel(const double *S, const int *D, uchar *C,
                               int width, int height);
